Added an 'Intersection' Handler module

+++ Code/Movement/TranslativeStepper.cpp
@@ -12,7 +12,7 @@
 //The difference between ISR Frequency and Calculation Frequency
 //The microstepping of the motor (assumed that it is a standard stepper motor with 200 steps per revolution)
 //The radius of the wheel used, the rotation of the wheel relative to the X-Axis of the robot, and the distance of the wheel to the center.
-TranslativeStepper::TranslativeStepper(volatile uint8_t *PORT, uint8_t pins, uint8_t ISRPerCal, uint8_t microstepping, float radius, float rotation, float distance) {
+TranslativeStepper::TranslativeStepper(volatile uint8_t *PORT, uint8_t pins, float ISRPerCal, uint8_t microstepping, float radius, float rotation, float distance) {
 	//Save these values
 	this->PORT = PORT;
 	this->pin = pins;
